pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "tph"
version = "0"
authors = [
  { name="Mike Cooke", email="drmikecooke@gmail.com" },
]
description = "Ppackage to download temperature, pressure, humidty from microbit."
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies =["pyserial"]

[tool.setuptools]
# ...
# By default, include-package-data is true in pyproject.toml, so you do
# NOT have to specify this line.
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
tphd = "tph.download:fileTPH"
tphcsv="tph.humanise:csvBuild"
